Trees Distributed Evenly
Because of the translateZ function, the parallax layers wasn't fulfilling the width and height of the windows (they were reduced), so the trees wasn't distributed evenly. To solve this, the property inset is added to each layer and the trees now looks for the layer width and height instead of looking for the windows width and height.